feat: migrated models to a models dir, same for tests
