+++ src/random_tree_models/models/gradientboostedtrees.py
@@ -13,7 +13,9 @@ from sklearn.utils.validation import (
     validate_data,  # type: ignore
 )
 
-import random_tree_models.decisiontree as dtree
+from random_tree_models.models.decisiontree import (
+    DecisionTreeRegressor,
+)
 from random_tree_models.params import MetricNames, is_greater_zero
 from random_tree_models.utils import bool_to_float
 
@@ -94,7 +96,7 @@ class GradientBoostedTreesRegressor(
     def fit(self, X: np.ndarray, y: np.ndarray) -> "GradientBoostedTreesRegressor":
         X, y = validate_data(self, X, y, ensure_all_finite=False)
 
-        self.trees_: T.List[dtree.DecisionTreeRegressor] = []
+        self.trees_: list[DecisionTreeRegressor] = []
 
         self.start_estimate_ = np.mean(y)
 
@@ -103,7 +105,7 @@ class GradientBoostedTreesRegressor(
 
         for _ in track(range(self.n_trees), total=self.n_trees, description="tree"):
             # train decision tree to predict differences
-            new_tree = dtree.DecisionTreeRegressor(
+            new_tree = DecisionTreeRegressor(
                 measure_name=self.measure_name,
                 max_depth=self.max_depth,
                 min_improvement=self.min_improvement,
@@ -240,7 +242,7 @@ class GradientBoostedTreesClassifier(
             raise ValueError("Cannot train with only one class present")
 
         self.classes_, y = np.unique(y, return_inverse=True)
-        self.trees_: T.List[dtree.DecisionTreeRegressor] = []
+        self.trees_: list[DecisionTreeRegressor] = []
         self.gammas_ = []
 
         y = self._bool_to_float(y)
@@ -254,7 +256,7 @@ class GradientBoostedTreesClassifier(
                 2 * y / (1 + np.exp(2 * y * yhat))
             )  # dloss/dyhat, g in the xgboost paper
 
-            new_tree = dtree.DecisionTreeRegressor(
+            new_tree = DecisionTreeRegressor(
                 measure_name=self.measure_name,
                 max_depth=self.max_depth,
                 min_improvement=self.min_improvement,
